requirements.txt
mcp>=1.0.0
openai>=1.0.0
anthropic>=0.7.0
pydantic>=2.0.0
python-dotenv>=1.0.0
asyncio-mqtt>=0.16.0
aiohttp>=3.8.0
typing-extensions>=4.0.0 